refactor: Disposal, Order 패키지에 리팩토링
