+++ src/main/java/uos/uos25/order/dto/response/DisposalCreateResponseDTO.java
@@ -1,37 +1,29 @@
-package uos.uos25.disposal.dto.response;
-
-import java.time.LocalDateTime;
+package uos.uos25.order.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
-import uos.uos25.disposal.entity.Disposal;
-import uos.uos25.product.dto.ProductInfo;
+import uos.uos25.headQuarter.dto.ProductInfo;
+import uos.uos25.order.entity.Disposal;
 import uos.uos25.shop.dto.ShopInfo;
 
 @Data
-public class DisposalGetResponseDTO {
+public class DisposalCreateResponseDTO {
     private final Long disposalId;
     private final Integer ea;
     private final ShopInfo shopInfo;
     private final ProductInfo productInfo;
-    private final LocalDateTime createAt;
 
     @Builder
-    public DisposalGetResponseDTO(
-            Long disposalId,
-            Integer ea,
-            ShopInfo shopInfo,
-            ProductInfo productInfo,
-            LocalDateTime createAt) {
+    public DisposalCreateResponseDTO(
+            Long disposalId, Integer ea, ShopInfo shopInfo, ProductInfo productInfo) {
         this.disposalId = disposalId;
         this.ea = ea;
         this.shopInfo = shopInfo;
         this.productInfo = productInfo;
-        this.createAt = createAt;
     }
 
-    public static DisposalGetResponseDTO fromEntity(Disposal disposal) {
-        return DisposalGetResponseDTO.builder()
+    public static DisposalCreateResponseDTO fromEntity(Disposal disposal) {
+        return DisposalCreateResponseDTO.builder()
                 .disposalId(disposal.getDisposalId())
                 .ea(disposal.getEa())
                 .shopInfo(
@@ -41,7 +33,6 @@ public class DisposalGetResponseDTO {
                         new ProductInfo(
                                 disposal.getProduct().getBarcode(),
                                 disposal.getProduct().getProductName()))
-                .createAt(disposal.getCreatedAt())
                 .build();
     }
 }
