+++ src/main/java/uos/uos25/order/dto/response/DisposalGetResponseDTO.java
@@ -1,29 +1,37 @@
-package uos.uos25.disposal.dto.response;
+package uos.uos25.order.dto.response;
+
+import java.time.LocalDateTime;
 
 import lombok.Builder;
 import lombok.Data;
-import uos.uos25.disposal.entity.Disposal;
-import uos.uos25.product.dto.ProductInfo;
+import uos.uos25.headQuarter.dto.ProductInfo;
+import uos.uos25.order.entity.Disposal;
 import uos.uos25.shop.dto.ShopInfo;
 
 @Data
-public class DisposalCreateResponseDTO {
+public class DisposalGetResponseDTO {
     private final Long disposalId;
     private final Integer ea;
     private final ShopInfo shopInfo;
     private final ProductInfo productInfo;
+    private final LocalDateTime createAt;
 
     @Builder
-    public DisposalCreateResponseDTO(
-            Long disposalId, Integer ea, ShopInfo shopInfo, ProductInfo productInfo) {
+    public DisposalGetResponseDTO(
+            Long disposalId,
+            Integer ea,
+            ShopInfo shopInfo,
+            ProductInfo productInfo,
+            LocalDateTime createAt) {
         this.disposalId = disposalId;
         this.ea = ea;
         this.shopInfo = shopInfo;
         this.productInfo = productInfo;
+        this.createAt = createAt;
     }
 
-    public static DisposalCreateResponseDTO fromEntity(Disposal disposal) {
-        return DisposalCreateResponseDTO.builder()
+    public static DisposalGetResponseDTO fromEntity(Disposal disposal) {
+        return DisposalGetResponseDTO.builder()
                 .disposalId(disposal.getDisposalId())
                 .ea(disposal.getEa())
                 .shopInfo(
@@ -33,6 +41,7 @@ public class DisposalCreateResponseDTO {
                         new ProductInfo(
                                 disposal.getProduct().getBarcode(),
                                 disposal.getProduct().getProductName()))
+                .createAt(disposal.getCreatedAt())
                 .build();
     }
 }
