+++ src/main/java/uos/uos25/order/repository/DisposalRepository.java
@@ -1,4 +1,4 @@
-package uos.uos25.disposal.repository;
+package uos.uos25.order.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.disposal.entity.Disposal;
+import uos.uos25.order.entity.Disposal;
 
 @Repository
 public interface DisposalRepository extends JpaRepository<Disposal, Long> {
